% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingsum.R
\name{missingsum}
\alias{missingsum}
\title{Missing Data Summary}
\usage{
missingsum(data, missing_values = c("", "NA", "N/A", "Missing"))
}
\arguments{
\item{data}{A data frame to analyze}

\item{missing_values}{A vector of additional values to treat as missing.
Can include both character and numeric values. Defaults to `c("", "NA", "N/A", "Missing")`.}
}
\value{
A nicely formatted table summarizing the missing data in the data frame.
  Includes the column names, count of missing values, and percentage of missing values.
}
\description{
This function provides a summary of missing data for each column in a data frame.
It identifies missing values based on `NA`, a user-specified set of placeholders
(e.g., `""`, `"N/A"`, `"Missing"` for characters or `-999`, `0` for numerics),
and returns a nicely formatted table.
}
\examples{
library(dplyr)
example_data <- data.frame(
  A = c(1, 2, NA, 4, ""),
  B = c("apple", "N/A", "banana", NA, "Missing"),
  C = c(10, 20, 30, 40, 50)
)
#using user-specified missing values
missingsum(example_data, missing_values = c("", "N/A", "Unknown"))

#using default missing values
missingsum(example_data)
}
